move route folder to server folder
