+++ server/config/app.js
@@ -18,8 +18,8 @@ liveReloadServer.server.once("connection", () => {
 });
 
 
-var indexRouter = require('../../routes/index');
-var usersRouter = require('../../routes/users');
+var indexRouter = require('../routes/index');
+var usersRouter = require('../routes/users');
 
 var app = express();
 app.use(connectLivereload());
